Reorganize file structure for inCore home page
