+++ src/app/incore/(home)/CaseStudiesSection.tsx
