+++ src/app/incore/(home)/DecorativeBannerSection.tsx
