+++ src/app/incore/(home)/IncoreINDsightsSection.tsx
