+++ src/app/incore/(home)/WealthSection.tsx
