+++ src/app/incore/(home)/page.tsx
@@ -4,14 +4,14 @@ import SectionHeader from "@/components/section-header";
 import { cn } from "@/lib/utils";
 import Image from "next/image";
 import Link from "next/link";
-import { ContactForm } from "../components/ContactForm";
-import CaseStudiesSection from "./components/CaseStudiesSection";
-import ClientsMarqueeSection from "./components/ClientsMarqueeSection";
-import DecorativeBannerSection from "./components/DecorativeBannerSection";
-import IncoreINDsightsSection from "./components/IncoreINDsightsSection";
-import ProcessJourneySection from "./components/ProcessJourneySection";
-import Testimonials from "./components/Testimonials";
-import WealthSection from "./components/WealthSection";
+import { ContactForm } from "../../components/ContactForm";
+import CaseStudiesSection from "./CaseStudiesSection";
+import ClientsMarqueeSection from "./ClientsMarqueeSection";
+import DecorativeBannerSection from "./DecorativeBannerSection";
+import IncoreINDsightsSection from "./IncoreINDsightsSection";
+import ProcessJourneySection from "./ProcessJourneySection";
+import Testimonials from "./Testimonials";
+import WealthSection from "./WealthSection";
 
 export default function InCorePage() {
   return (
